% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as_dwc}
\alias{as_dwc}
\title{Modify an input OBIS dataset to include only Darwin Core required and 
recommended fields.}
\usage{
as_dwc(x, template = template_dwc(n = 1))
}
\arguments{
\item{x}{tibble of raw OBIS data}

\item{template}{template as a tibble}
}
\value{
tibble with defined columns
}
\description{
Modify an input OBIS dataset to include only Darwin Core required and 
recommended fields.
}
\seealso{
\href{https://ipt.gbif.org/manual/en/ipt/2.5/occurrence-data}{Darwin Core}
}
