% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fetch_obis}
\alias{fetch_obis}
\title{Fetch a species from OBIS}
\usage{
fetch_obis(
  scientificname = "Carcharodon carcharias",
  save_file = file_name(scientificname),
  template = species_template(),
  ...
)
}
\arguments{
\item{scientificname}{character, the Latin name for a species}

\item{save_file}{NA or a path specification to save the file}

\item{template}{data frame defining minimal fields}

\item{...}{other arguments for \code{\link[robis]{occurrence}}}
}
\value{
tibble, possibly empty if a species is not found
}
\description{
Fetch a species from OBIS
}
