% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{file_name}
\alias{file_name}
\title{Convert from scientific name to file name}
\usage{
file_name(
  x = "Carcharodon carcharias",
  ext = ".csv.gz",
  sep = "_",
  path = get_path()
)
}
\arguments{
\item{x}{chr, scientific name of species}

\item{ext}{chr, file extension to be added}

\item{sep}{chr, separator of file name (between genus and species)}

\item{path}{chr, path of which file is stored}
}
\value{
character vector of file names
}
\description{
Convert from scientific name to file name
}
