% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{get_path}
\alias{get_path}
\title{Retrieve a data path}
\usage{
get_path(..., root = rappdirs::user_data_dir("obis_niche"))
}
\arguments{
\item{...}{character, one or more file path segments to be post-pended to \code{root}}

\item{root}{character, the root data directory path}
}
\value{
character path specification
}
\description{
Retrieve a data path
}
