% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{list_obis}
\alias{list_obis}
\title{Retrieve a listing of local files}
\usage{
list_obis(path = get_path(), full.names = FALSE, strip = c(NA, ".csv.gz")[2])
}
\arguments{
\item{path}{the path to the datasets}

\item{full.names}{logical return basename or full path specs?}

\item{strip}{character or NA, if character then strip this pattern. Ignored
if full.names is \code{TRUE}}
}
\value{
character names (possibly filenames)
}
\description{
Retrieve a listing of local files
}
