% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obisniche_package.R
\docType{package}
\name{obisniche}
\alias{obisniche}
\title{obisniche: R tool for assessing uncertainties around qualities of observation for environmental niches using OBIS data.}
\description{
R tool for assessing uncertainties around qualities of observation for environmental niches using OBIS data.
}
