% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_obis}
\alias{plot_obis}
\title{Plot the locations of a OBIS dataset}
\usage{
plot_obis(x, what = c("base", "ggplot", "leaflet", "gist")[1], ...)
}
\arguments{
\item{x}{tibble of OBIS data}

\item{what}{character, one of 'base', 'ggplot', 'leaflet', 'gist'}

\item{...}{other arguments for mapr package map_* functions}
}
\description{
Plot the locations of a OBIS dataset
}
