% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{rename_obis}
\alias{rename_obis}
\title{Rename one or more fields (columns) in a OBIS tibble}
\usage{
rename_obis(
  x,
  fields = c(name = "scientificName", longitude = "decimalLongitude", latitude =
    "decimalLatitude")
)
}
\arguments{
\item{x}{tibble of data}

\item{fields}{named character vector with oldname = newname}
}
\value{
renamed tibble
}
\description{
Rename one or more fields (columns) in a OBIS tibble
}
