% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{template_dwc}
\alias{template_dwc}
\title{Create a template with required, recommended or both (default) columns.}
\usage{
template_dwc(what = c("required", "recommended"), n = 0)
}
\arguments{
\item{what}{character, one or more of 'required' and 'recommended'}

\item{n}{numeric, the number of rows to return,  less than equal to zero
yields an empty tibble}
}
\value{
tibble with n rows
}
\description{
Create a template with required, recommended or both (default) columns.
}
